feat(community-tabs): added tab panel in tabs component
