+++ packages/Tabs/Tabs.md
@@ -1,44 +1,21 @@
 ### Usage criteria
 
-- Use `Tabs` according to props data
+Use `AddTabs` according to props data
+Note that the `copy` prop must be provided at all times for the correct accessible text.
 
 ```jsx
-<Tabs
-  tabIndex={0}
-  tabsEntries={[
-    {
-      entryTitle: 'Modal - Channels - Tabs - Essentials copy',
-      tabTitle: 'Essentials',
-      tabContent: [
-        {
-          entryTitle: 'Essentials Modal - Channels - Tabs - Essentials - Row copy',
-          hasGutter: false,
-          horizontalAlign: 'start',
-        },
-      ],
-    },
-    {
-      entryTitle: 'Modal - Channels - Tabs - Essentials copy',
-      tabTitle: 'Premium',
-      tabContent: [
-        {
-          entryTitle: 'Premium modal - Channels - Tabs - Essentials - Row copy',
-          hasGutter: false,
-          horizontalAlign: 'start',
-        },
-      ],
-    },
-    {
-      entryTitle: 'Modal - channels - Tabs - Essentials copy',
-      tabTitle: 'Theme packs',
-      tabContent: [
-        {
-          entryTitle: 'Theme modal - Channels - Tabs - Essentials - Row copy',
-          hasGutter: false,
-          horizontalAlign: 'start',
-        },
-      ],
-    },
-  ]}
-/>
+<AddTabs copy="en" tabs={['Themepacks', 'Premium', 'Essentials']}>
+  <AddTabs.Panel>
+    Choose from more than 40 options and customize them any time in your online account or on the My
+    TELUS app.
+  </AddTabs.Panel>
+  <AddTabs.Panel>
+    Choose from more than 15 options and customize them any time in your online account or on the My
+    TELUS app.
+  </AddTabs.Panel>
+  <AddTabs.Panel>
+    A great selection of channels from major Canadian and U.S. networks, plus get 72 Stingray music
+    channels.
+  </AddTabs.Panel>
+</AddTabs>
 ```
