+++ packages/Tabs/__tests__/Tabs.spec.jsx
@@ -1,30 +1,31 @@
 import React from 'react'
-import { shallow } from 'enzyme'
+import { mount } from 'enzyme'
 
-import Tabs from '../Tabs'
+import AddTabs from '../Tabs'
 
-describe('Tabs', () => {
-  const doShallow = (props = {}) => shallow(<Tabs {...props} />)
+describe('AddTabs', () => {
+  const children = (
+    <AddTabs copy="en">
+      <AddTabs.Panel>Content 1</AddTabs.Panel>
+      <AddTabs.Panel>Content 2</AddTabs.Panel>
+      <AddTabs.Panel>Content 3</AddTabs.Panel>
+    </AddTabs>
+  )
+  const doMount = () => mount(children)
 
   it('renders', () => {
-    const tabs = doShallow()
-
+    const tabs = doMount()
     expect(tabs).toMatchSnapshot()
   })
-
-  it('passes additional attributes to the element', () => {
-    const tabs = doShallow({ id: 'the-id', 'data-some-attr': 'some value' })
-
-    expect(tabs).toHaveProp('id', 'the-id')
-    expect(tabs).toHaveProp('data-some-attr', 'some value')
+  it('does other things', () => {
+    const tabs = doMount()
+    expect(tabs).toExist()
   })
-
   it('does not allow custom CSS', () => {
-    const tabs = doShallow({
+    const tabs = doMount({
       className: 'my-custom-class',
       style: { color: 'hotpink' },
     })
-
     expect(tabs).not.toHaveProp('className', 'my-custom-class')
     expect(tabs).not.toHaveProp('style')
   })
