+++ packages/Tabs/package.json
@@ -28,6 +28,11 @@
   "dependencies": {
     "@tds/util-helpers": "^1.2.0",
     "prop-types": "^15.6.2",
-    "react-tabs": "^3.1.1"
+    "react-tabs": "^3.1.1",
+    "@tds/core-a11y-content": "^2.1.5",
+    "@tds/core-colours": "^2.2.1",
+    "@tds/core-decorative-icon": "^2.6.4",
+    "@tds/core-responsive": "^1.3.3",
+    "string-hash": "^1.1.3"
   }
 }
